third_party/blink/web_tests/jquery/deferred-expected.txt

ALERT: 52 tests of 52 passed, 0 failed.
